Add fail action to the progress bar
