+++ src/main.ts
@@ -5,21 +5,6 @@ import store from '@/store'
 import http from '@/http'
 import i18n from '@/i18n'
 import '@/registerServiceWorker'
-const VueProgressBar = require('vue-progressbar')
-
-const options = {
-  color: '#f4b9b9',
-  failedColor: '#b22020',
-  thickness: '5px',
-  transition: {
-    speed: '0.1s',
-    opacity: '0.3s',
-    termination: 300
-  },
-  autoRevert: true,
-  location: 'top'
-}
-Vue.use(VueProgressBar, options)
 
 Vue.prototype.$http = http
