+++ src/views/helper.js
@@ -1,5 +1,6 @@
 import i18n from '@/i18n'
 import store from '@/store'
+import progressBar from '@/progressBar'
 
 export const fetchWithProgress = async (targets, errText) => {
   store.dispatch('fetch/startProgress')
@@ -16,7 +17,7 @@ export const fetchWithProgress = async (targets, errText) => {
         text: i18n.t(errText) + (err.apierr ? '\n' + err.apierr : '')
       })
     }
-
+    progressBar.$Progress.fail()
     store.dispatch('fetch/endProgress')
 
     throw err
